msvc build fixes
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@8080 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/switch_ivr_play_say.c
@@ -1776,7 +1776,7 @@ SWITCH_DECLARE(switch_status_t) switch_ivr_soft_hold(switch_core_session_t *sess
 	switch_input_args_t args = { 0 };
 	args.input_callback = hold_on_dtmf;
 	args.buf = (void *) unhold_key;
-	args.buflen = strlen(unhold_key);
+	args.buflen = (uint32_t)strlen(unhold_key);
 	
 	switch_assert(session != NULL);
 	channel = switch_core_session_get_channel(session);
